break GUI into separate files
