+++ src/main/python/main.py
@@ -1,11 +1,7 @@
 from PyQt5 import QtWidgets
 from fbs_runtime.application_context.PyQt5 import ApplicationContext
 import sys
-from ntu_learn_downloader_gui.gui import (
-    LoginDialog,
-    ChooseDirDialog,
-    DownloadDialog
-)
+from ntu_learn_downloader_gui.gui.login_dialog import LoginDialog
 
 if __name__ == "__main__":
     appctxt = ApplicationContext()
